Refactor updateMTOReviewedBillableWeightsAt
