+++ src/constants/MoveHistory/EventTemplates/UpdateMTOReviewedBillableWeightAt/updateMTOReviewedBillableWeightsAt.jsx
@@ -1,4 +1,5 @@
-import detailsTypes from 'constants/MoveHistory/UIDisplay/DetailsTypes';
+import React from 'react';
+
 import operations from 'constants/MoveHistory/UIDisplay/Operations';
 import actions from 'constants/MoveHistory/Database/Actions';
 import tables from 'constants/MoveHistory/Database/Tables';
@@ -7,7 +8,6 @@ export default {
   action: actions.UPDATE,
   eventName: operations.updateMTOReviewedBillableWeightsAt,
   tableName: tables.moves,
-  detailsType: detailsTypes.PLAIN_TEXT,
   getEventNameDisplay: () => 'Updated move',
-  getDetailsPlainText: () => 'Reviewed weights',
+  getDetails: () => <> Reviewed weights </>,
 };
